Fixed minor mistakes
Now the CGMid functions have all of the bytes, the byte array initializes the 12 bytes for the offsets. Also the default Effect instruction has its argument parsed correctly.

With the stock JSON obtained from the game, a 1:1 replica has been obtained, so we can conclude that the decompilation has been completed.